BUG: reset attribute `tla.lex.Lexer.lexer.lineno`
and also `string_start` (not initializing and
not resetting `string_start` is not an issue).

Previously, tokenizing a sequence of input strings
resulted in the starting line of the second
string being the last line of the previous string.

This error affected only lexing by directly
instantiating the class `tla.lex.Lexer`.
This error did not affect the functions:

- `tla.lex.tokenize`
- `tla.lex.lex`
- `tla.parser.parse`
- `tla.parser.parse_expr`

because these create a new lexer instance
each time they are called.

That `ply` does not reset the attribute `lineno`
of the lexer is noted in the section ["EOF Handling"](
    https://www.dabeaz.com/ply/ply.html#ply_nn14)
of the documentation of `ply`, and in the section
["Internal lexer state"](
    https://www.dabeaz.com/ply/ply.html#ply_nn20).