TST: also same expressions delimited by parentheses
because currently, the function `tla.parse_expr` parses
up to where it can, and leaves the rest of the input
unprocessed. This API may change in the future.
Until that happens, these newly added tests detect
unconsumed input when calling the function `parse_expr`.

These tests are useful in any case (i.e., after changes to
the API), because they test parsing of parenthesized expressions.